Moves: source.go -> path.go
