+++ transform/path.go
@@ -13,6 +13,41 @@ import (
 var InvalidCNPJError = errors.New("Invalid CNPJ")
 var InvalidPathError = errors.New("Invalid path for a CNPJ JSON file")
 
+type sourceType string
+
+const (
+	venue         sourceType = "ESTABELE"
+	motive                   = "MOTICSV"
+	main                     = "EMPRECSV"
+	city                     = "MUNICCSV"
+	cnae                     = "CNAECSV"
+	country                  = "PAISCSV"
+	nature                   = "NATJUCSV"
+	partner                  = "SOCIOCSV"
+	qualification            = "QUALSCSV"
+	simple                   = "SIMPLES"
+)
+
+func PathsForSource(t sourceType, dir string) ([]string, error) {
+	var ls []string
+
+	r, err := os.ReadDir(dir)
+	if err != nil {
+		return ls, err
+	}
+
+	s := strings.ToLower(string(t) + ".zip")
+	for _, f := range r {
+		if f.IsDir() {
+			continue
+		}
+		if strings.HasSuffix(strings.ToLower(f.Name()), s) {
+			ls = append(ls, filepath.Join(dir, f.Name()))
+		}
+	}
+	return ls, nil
+}
+
 func PathForCNPJ(c string) (string, error) {
 	if !cnpj.IsValid(c) {
 		return "", fmt.Errorf("Error finding file path for %s: %w", c, InvalidCNPJError)
